LayoutTests/media/video-dom-start.html

<video src=content/test.mp4></video>
<script src=video-test.js></script>
<script>
video.start = 1.0;
test("video.start == 1.0");
test("video.getAttribute('start') == '1s'");
waitForEventTestAndEnd('canplaythrough', "video.currentTime == 1.0");
</script>
