LayoutTests/media/video-dom-start-expected.txt

TEST(video.start == 1.0) OK
TEST(video.getAttribute('start') == '1s') OK
EVENT(canplaythrough) TEST(video.currentTime == 1.0) OK
END OF TEST
